Move AppRoot to own directory and rename as AppLayout
