+++ saleor/static/dashboard-next/components/AppLayout/consts.ts
@@ -1,0 +1,2 @@
+export const drawerWidth = 256;
+export const navigationBarHeight = 64;
