+++ saleor/static/dashboard-next/components/AppLayout/index.ts
@@ -1,0 +1,2 @@
+export { default } from './AppLayout';
+export * from './AppLayout';
